+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.1
+
+* Platform permission supports
+* Auto storage permission detecting on platform
+
 ## 1.1.0
 
 * Permission handler supports
+
+## 1.0.0
+
+* Initial release